dam-mcp/cloud_functions/granola_sync/requirements.txt
functions-framework==3.*
google-cloud-firestore>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
pyyaml>=6.0.0
